Fix the following issues:
Some of my testing songs don't have complete info. For example, albumName is nil. It causes the app to crash when trying to assemble the dictionary for the song.

The background event handler considers only the RemoteControlTogglePlayPause event but on my phone (iPhone 4, iOS 7), tapping the pause button on the lock screen generates a RemoteControlPause event (similar issue with the play button).

After I fix issue 2, tapping the pause button does paused the music, but the playback lock screen (with playback controls) disappears, reverting back to the lock screen with wallpaper. This can be fixed by setting the nowPlayingInfo.